+++ official/vision/detection/models/retinanet.py
@@ -33,7 +33,7 @@ class RetinaNet(M.Module):
         )
         self.box_coder = layers.BoxCoder(cfg.reg_mean, cfg.reg_std)
 
-        self.stride_list = np.array(cfg.stride).astype(np.float32)
+        self.stride_list = np.array(cfg.stride, dtype=np.float32)
         self.in_features = ["p3", "p4", "p5", "p6", "p7"]
 
         # ----------------------- build the backbone ------------------------ #
@@ -82,8 +82,8 @@ class RetinaNet(M.Module):
 
     def preprocess_image(self, image):
         normed_image = (
-            image - np.array(self.cfg.img_mean)[None, :, None, None]
-        ) / np.array(self.cfg.img_std)[None, :, None, None]
+            image - np.array(self.cfg.img_mean, dtype=np.float32)[None, :, None, None]
+        ) / np.array(self.cfg.img_std, dtype=np.float32)[None, :, None, None]
         return layers.get_padded_tensor(normed_image, 32, 0.0)
 
     def forward(self, inputs):
